added bids service and products service and controllers

+++ src/main/java/com/intuit/bidding/service/repository/UserInfoRepository.java
@@ -8,7 +8,7 @@ import org.springframework.stereotype.Repository;
 import java.util.Optional;
 
 @Repository
-public interface UserInfoRepository extends JpaRepository<UserInfo,Integer> {
+public interface UserInfoRepository extends JpaRepository<UserInfo,Long> {
 
     Optional<UserInfo> findByName(String userName);
